correctionOfErrors

+++ error.cpp
@@ -8,12 +8,12 @@ Error::~Error(){
     
 }  
 
-double Error::calculateError(std::vector<int> &received, Word &original){
+double Error::calculateError(std::vector<int>& received, Word& original){
     double error = 0.0;
     for(int i = 0; i < received.size(); i++){
         if(received[i] != original.get(i)){
             error++;
         }
     }
-    return error/received.size();
+    return error/(double)received.size();
 }
